--- Backend/node-postgres/src/routes/products.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const product= require('../controllers/product.controller');
-//console.log(product)
-
-// /
-router.post('/',product.createProduct);
-router.get('/',product.getProducts);
-
-// /:productID
-router.get('/:id',product.getOneProduct);
-router.put('/:id',product.updateProduct);
-router.delete('/:id',product.deleteProduct);
-module.exports = router;
